chore: 중복 정의된 타입 제가, 정의된 타입 사용
