+++ src/app/components/card/cardList/MyApplicationListItem.tsx
@@ -2,7 +2,7 @@ import { getRecruitmentStatus } from "@/utils/recruitDateFormatter";
 import { formatLocalDate } from "@/utils/workDayFormatter";
 import Chip from "@/app/components/chip/Chip";
 import Image from "next/image";
-import { ApplyStatus } from "@/types/application";
+import { ApplicationStatusType } from "@/types/application";
 import axios from "axios";
 import toast from "react-hot-toast";
 import { MyApplicationType } from "@/types/response/user";
@@ -88,8 +88,8 @@ const MyApplicationListItem = ({ createdAt, status, resumeId, resumeName, form }
         <div className="text-grayscale-700 mt-4 flex h-[50px] items-center justify-start gap-2 rounded-2xl text-sm lg:text-base">
           <div className="rounded-[4px] border border-primary-orange-300 bg-primary-orange-50">
             <Chip
-              label={translateStatus(status as ApplyStatus)}
-              variant={getStatusVariant(status as ApplyStatus)}
+              label={translateStatus(status as ApplicationStatusType)}
+              variant={getStatusVariant(status as ApplicationStatusType)}
               textStyle="font-bold"
             />
           </div>
