+++ src/app/components/modal/modals/confirm/SelectProgressModal.tsx
@@ -3,16 +3,16 @@ import { cn } from "@/lib/tailwindUtil";
 import RadioBtn from "@/app/components/button/default/RadioBtn";
 import RadioGroup from "@/app/components/button/default/RadioGroup";
 import Button from "@/app/components/button/default/Button";
-import { ApplyStatus, applicationStatus } from "@/types/application";
 import { positionOptions } from "@/constants/positionOptions";
 import axios from "axios";
 import toast from "react-hot-toast";
 import type { ConfirmFormModalProps } from "@/types/modal";
 import DotLoadingSpinner from "@/app/components/loading-spinner/DotLoadingSpinner";
+import { applicationStatus, ApplicationStatusType } from "@/types/applicationStatus";
 
 const SelectProgressModal = ({ id, isOpen, onClose, className }: ConfirmFormModalProps) => {
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const [selectedValue, setSelectedValue] = useState<ApplyStatus>(applicationStatus.INTERVIEW_PENDING);
+  const [selectedValue, setSelectedValue] = useState<ApplicationStatusType>(applicationStatus.INTERVIEW_PENDING);
 
   if (!isOpen) return null;
 
@@ -23,7 +23,7 @@ const SelectProgressModal = ({ id, isOpen, onClose, className }: ConfirmFormModa
     { id: "hired", value: applicationStatus.HIRED, label: "채용 완료" },
   ] as const;
 
-  const handleValueChange = (value: ApplyStatus) => {
+  const handleValueChange = (value: ApplicationStatusType) => {
     setSelectedValue(value);
   };
